td-shim-tools: merge td-shim-sign-payload into td-shim-tools
Merge td-shim-sign-payload into td-shim-tools.
