+++ doc/secure_boot_guide.md
@@ -61,7 +61,7 @@ set AR=
 
 Run the signing tool:
 ```
-cargo run -p td-shim-tools --bin td-shim-sign-payload -- -A ECDSA_NIST_P384_SHA384 target/x86_64-unknown-none/release/td-payload 1 1 data/sample-keys/ecdsa-p384-private.pk8 
+cargo run -p td-shim-tools --bin td-shim-sign-payload -- -A ECDSA_NIST_P384_SHA384 data/sample-keys/ecdsa-p384-private.pk8 target/x86_64-unknown-none/release/td-payload 1 1 
 ```
 The signed payload file **td-payload-signed** is located in the same folder with input `td-payload`.
