--- td-shim-sign-payload/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "td-shim-sign-payload"
-version = "0.1.0"
-description = "Sign shim payload binary with RSA/DSA"
-repository = "https://github.com/confidential-containers/td-shim"
-homepage = "https://github.com/confidential-containers"
-license = "BSD-2-Clause-Patent"
-edition = "2018"
-
-[dependencies]
-r-efi = "3.2.0"
-scroll = { version = "0.10", default-features = false, features = ["derive"]}
-
-clap = { version = "3.0", features = ["cargo"], optional = true }
-der = { version = "0.4.5", features = ["oid"], optional = true}
-env_logger = { version = "0.9.0", optional = true }
-log = { version = "0.4.5", optional = true }
-ring = { git = "https://github.com/jyao1/ring.git", branch = "uefi_support", default-features = false, features = ["alloc"], optional = true}
-td-layout = { path = "../td-layout", optional = true }
-td-shim-ld = { path = "../td-shim-ld", default-features = false, optional = true}
-
-[features]
-default = ["signer"]
-signer = ["clap", "der", "env_logger", "log", "ring", "td-layout", "td-shim-ld/linker"]
